C'est la pls mais ça marche

+++ NET_Project/DLL_Library_NET_Project/Business/Recette.cs
@@ -4,7 +4,7 @@
     {
         public int id { get; set; }
         public string intitule { get; set; }
-        public double prix { get; set; }
+        public int prix { get; set; }
         public int temps_prepa { get; set; }
         public virtual TypeRecette TypeRecette { get; set; }
     }
